requirements.txt
dj-database-url==0.5.0
Django==2.2.1
django-ckeditor==5.7.1
django-crispy-forms==1.7.2
django-js-asset==1.2.2
gunicorn==19.9.0
psycopg2==2.8.2
pytz==2019.1
sqlparse==0.3.0
whitenoise==4.1.2
